Add soft dependencies
These continue to be satisfied after a task is done, so long as it is
not reset back to a pending state, even if it is cleaned.

We're also sneaking in some race condition prevention by ensuring that
task_dependency stays locked until the end of the transaction. This
prevents other transactions from looking at it and not seeing the new
row.