Fix count of satisfied dependencies
Previously, task.num_dependencies_incomplete was intended to count the
number of dependencies that are in an incomplete state (i.e. have
task_state.incomplete). It didn't do this properly, because when a new
dependency was added, the trigger would check the values of
task_state.satisfies_dependency and task_state.satisfies_soft_dependency
instead, and update the counter to reflect the number of satisfied
dependencies. This is actually the desired behaviour, so the discrepancy
could've gone unnoticed, except that the trigger that runs when the
state of a task changes was still performing its contractually-obligated
duties. This meant that task.num_dependencies_incomplete wasn't actually
tracking any single coherent property.

However, that didn't matter in most cases, because the only time the
distinction came into play was when cleaned tasks (which could be
complete and yet not satisfy dependencies) had pending dependents. In
regular use, this situation could arise when rerunning a task, which is
not supposed to happen very often.

The new column task.num_dependencies_satisfied is supposed to count the
number of dependencies (soft or otherwise) that are currently satisfied.
This makes it useful when deciding whether a task is ready to be run,
and its name properly reflects the intent behind it.